multitimer/sched: use triggercheck for pre-alarm code
