+++ apps/multitimer/boot.js
@@ -1,0 +1,8 @@
+{
+	const resetTimer = alarm => {
+		if (alarm.timer) alarm.timer = alarm.data.ot;
+	};
+
+	Bangle.on("alarmSnooze", resetTimer);
+	Bangle.on("alarmDismiss", resetTimer);
+}
